api/requirements.txt
fastapi[standard]
pydantic
pymongo
python-dotenv
bcrypt
cryptography
python-jose
python-multipart
passlib

